connect message unmarshalling
